pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "nanohorizon"
version = "0.1.0"
description = "Crafter ML engineering golf scaffolding"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.28.1",
  "pyyaml>=6.0.2",
]

[tool.hatch.build.targets.wheel]
packages = ["src/nanohorizon"]

[dependency-groups]
dev = [
  "pre-commit>=4.0.0",
  "pytest>=8.4.2",
  "ruff>=0.9.0",
  "ty>=0.0.20",
]
modal = [
  "modal>=1.0.0",
]
cloud = [
  "synth-ai @ file:///Users/joshpurtell/Documents/GitHub/synth-ai",
]
training = [
  "accelerate>=1.10.0",
  "datasets>=4.1.0",
  "peft>=0.15.0",
  "torch>=2.6.0",
  "transformers @ git+https://github.com/huggingface/transformers.git@main",
  "trl>=0.28.0",
  "vllm>=0.10.0",
]

[tool.uv]
default-groups = ["dev"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]
extend-exclude = [
  "artifacts",
  ".out",
  "containers/crafter_rs/target",
]

[tool.ruff.lint]
select = [
  "E",   # pycodestyle errors
  "W",   # pycodestyle warnings
  "F",   # pyflakes
  "I",   # isort
  "UP",  # pyupgrade
  "B",   # flake8-bugbear
  "SIM", # flake8-simplify
]
ignore = [
  "E501", # line length (formatter wraps where it can)
]

[tool.ruff.lint.isort]
known-first-party = ["nanohorizon"]

[tool.ruff.format]
quote-style = "double"

[tool.ty.environment]
root = ["./src"]
python-version = "3.11"

[tool.ty.src]
include = ["src"]

[tool.ty.analysis]
# Training stack is optional (see `training` dependency group); default dev venv omits these.
replace-imports-with-any = [
  "torch.**",
  "transformers.**",
  "trl.**",
  "peft.**",
  "accelerate.**",
  "datasets.**",
  "vllm.**",
]
